feat(web): update password on pentester and manager settings
